+++ cypress/support/stubWithFixture.js
@@ -1,7 +1,3 @@
 Cypress.Commands.add('stubWithFixture', ({ method = 'GET', url, fixture }) => {
-    return cy.route({
-        method,
-        url,
-        response: `fixture:${fixture}`,
-    })
+    return cy.intercept(method, url, { fixture })
 })